+++ Proiect-TroubleShooters/Proiect-TroubleShooters/Player.cpp
@@ -16,11 +16,11 @@ void Player::displayPlayerNumberPieces()
 	std::cout << "Pieces: " << maxPieces - getNumberPieces() << '\n';
 	std::cout << "Bridges: " << maxPieces - getNumberBridges() << '\n';
 }
-const uint32_t& Player::getNumberMaxPieces() const
+const uint16_t& Player::getNumberMaxPieces() const
 {
 	return maxPieces;
 }
-const uint32_t& Player::getScore() const
+const uint16_t& Player::getScore() const
 {
 	return score;
 }
@@ -45,7 +45,7 @@ std::vector<Bridge>& Player::getBridges()
 	return bridges;
 }
 
-void Player::setMaxPieces(const uint32_t& maxPieces)
+void Player::setMaxPieces(const uint16_t& maxPieces)
 {
 	this->maxPieces = maxPieces;
 }
